Free query resources after use
Introduce a macro WITH-PREPARED-QUERY. Given a connection and sql statement, it
prepares a statement, binds it to desired name, executes the body, and then
cleans up the query with an UNWIND-PROTECT.

Replaced all calls to DBI:PREPARE with the equivalent WITH-PREPARED-BODY.

This helps prevent unused resources from piling up. As an example, cl-sqlite
finalizes all queries when the database connection is closed. In a program that
connects, performs millions of queries, and then disconnects, the disconnect
will take a ridiculous amount of time to perform if the resources are not freed
immediately after use.