Fixed error when insert-dao used with object having the value for the serial-key.
I encounter this error from PostgreSQL:

    2020-11-08 15:03:44.000 UTC [30315] ERROR:  currval of sequence "dist_id_seq" is not yet defined in this session
    2020-11-08 15:03:44.000 UTC [30315] STATEMENT:  SELECT currval(pg_get_serial_sequence('dist', 'id')) AS last_insert_id
    2020-11-08 15:03:44.001 UTC [30315] ERROR:  current transaction is aborted, commands ignored until end of transaction block

When trying to pass an object with `id` into the insert-dao.

This is a rare situation because I'm having a composite primary key on my object: `(id version)`

ID field is autoincremented, but version is not.

When I create a new object version, I call `insert-dao` on the object where ID slot has value and VERSION slot
is incremented by application code. This leads to the error, because Mito tries to fetch last insert id from the sequence.

This commit fixes this problem by not trying to fetch last insert id if its value already known.